Add workflow to auto-import Weblate changes
If the Weblate GitHub user opens a PR, this will import the Weblate translations in that PR into the prelude, then export them, then create a diff of the changes that is used to add a commit to the PR's branch containing those changes.

The export is done because Weblate just appends new translations to its YAML files, but they're otherwise sorted alphabetically and the export is a way of doing that.